// apps/web/src/lib/api-client/models/taskDtoPriority.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * AppStart API
 * API REST do AppStart - Especificação OpenAPI 3.0
 * OpenAPI spec version: 1.0.0
 */

/**
 * Prioridade da tarefa
 */
export type TaskDtoPriority = typeof TaskDtoPriority[keyof typeof TaskDtoPriority];


export const TaskDtoPriority = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
  URGENT: 'URGENT',
} as const;