// apps/web/src/lib/api-client/models/taskDtoStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * AppStart API
 * API REST do AppStart - Especificação OpenAPI 3.0
 * OpenAPI spec version: 1.0.0
 */

/**
 * Status da tarefa
 */
export type TaskDtoStatus = typeof TaskDtoStatus[keyof typeof TaskDtoStatus];


export const TaskDtoStatus = {
  PENDING: 'PENDING',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
} as const;